requirements.txt
anthropic
openai
atlassian-python-api
sqlalchemy
flask
apscheduler
pyyaml
sentence_transformers
pinecone-client
weaviate-client
python-dotenv
requests
tzdata
dotenv
sqlalchemy
pyodbc
anthropic
chromadb
sentence_transformers
jira
boto3